+++ src/exchanges/bitfinex/modules/authed/position/close.ts
@@ -10,7 +10,7 @@ import {
 } from '../../../../../lib/modules/authed/IAlunaPositionModule'
 import { IAlunaPositionSchema } from '../../../../../lib/schemas/IAlunaPositionSchema'
 import { BitfinexHttp } from '../../../BitfinexHttp'
-import { placeMarketOrderToClosePosition } from './helpers/placeMarketOrderToClosePosition'
+import { placeMarketOrderToClosePosition } from '../../../../../utils/positions/placeMarketOrderToClosePosition'
 import { throwPositionIdRequiredFor } from './helpers/throwPositionIdRequiredFor'
 
 
@@ -47,7 +47,7 @@ export const close = (exchange: IAlunaExchangeAuthed) => async (
 
   }
 
-  await placeMarketOrderToClosePosition({
+  const { order } = await placeMarketOrderToClosePosition({
     http,
     exchange,
     position,
@@ -57,6 +57,7 @@ export const close = (exchange: IAlunaExchangeAuthed) => async (
     ...position,
     status: AlunaPositionStatusEnum.CLOSED,
     closedAt: new Date(),
+    closePrice: order.rate!,
   }
 
   const { requestWeight } = http