+++ src/exchanges/ftx/modules/authed/order/listRaw.spec.ts
@@ -1,0 +1,54 @@
+import { expect } from 'chai'
+
+import { mockHttp } from '../../../../../../test/mocks/exchange/Http'
+import { AlunaHttpVerbEnum } from '../../../../../lib/enums/AlunaHtttpVerbEnum'
+import { IAlunaCredentialsSchema } from '../../../../../lib/schemas/IAlunaCredentialsSchema'
+import { FtxAuthed } from '../../../FtxAuthed'
+import { FtxHttp } from '../../../FtxHttp'
+import { getFtxEndpoints } from '../../../ftxSpecs'
+import { FTX_RAW_ORDERS } from '../../../test/fixtures/ftxOrders'
+
+
+
+describe(__filename, () => {
+
+  const credentials: IAlunaCredentialsSchema = {
+    key: 'key',
+    secret: 'secret',
+  }
+
+  it('should list Ftx raw orders just fine', async () => {
+
+    // preparing data
+    const mockedRawOrders = FTX_RAW_ORDERS
+
+    // mocking
+    const {
+      publicRequest,
+      authedRequest,
+    } = mockHttp({ classPrototype: FtxHttp.prototype })
+
+    authedRequest.returns(Promise.resolve(mockedRawOrders))
+
+    // executing
+    const exchange = new FtxAuthed({ credentials })
+
+    const { rawOrders } = await exchange.order.listRaw()
+
+
+    // validating
+    expect(rawOrders).to.deep.eq(mockedRawOrders)
+
+    expect(authedRequest.callCount).to.be.eq(1)
+
+    expect(authedRequest.firstCall.args[0]).to.deep.eq({
+      verb: AlunaHttpVerbEnum.GET,
+      credentials,
+      url: getFtxEndpoints(exchange.settings).order.list,
+    })
+
+    expect(publicRequest.callCount).to.be.eq(0)
+
+  })
+
+})