+++ src/exchanges/ftx/modules/public/symbol/list.ts
@@ -1,0 +1,35 @@
+import debug from 'debug'
+
+import { IAlunaExchangePublic } from '../../../../../lib/core/IAlunaExchange'
+import {
+  IAlunaSymbolListParams,
+  IAlunaSymbolListReturns,
+} from '../../../../../lib/modules/public/IAlunaSymbolModule'
+import { FtxHttp } from '../../../FtxHttp'
+
+
+
+const log = debug('alunajs:ftx/symbol/list')
+
+
+
+export const list = (exchange: IAlunaExchangePublic) => async (
+  params: IAlunaSymbolListParams = {},
+): Promise<IAlunaSymbolListReturns> => {
+
+  log('listing Ftx symbols')
+
+  const { http = new FtxHttp(exchange.settings) } = params
+
+  const { requestWeight } = http
+
+  const { rawSymbols } = await exchange.symbol.listRaw({ http })
+
+  const { symbols } = exchange.symbol.parseMany({ rawSymbols })
+
+  return {
+    symbols,
+    requestWeight,
+  }
+
+}